Massive Changes
Firstly I have created a React with TypeScript UI which handles barcode decoding and webcam access - compatible with arm64.
This allows for automatic barcode detection and decoding whilst running on a web browser.
I have also configured RestAPI to send the decoded barcode to the java backend - which calls the Open Food Facts API, returning a json of product information corresponding to the barcode which is then sent back in the response to the web server.